@@ -2,8 +2,8 @@ package com.tequila.ecommerce.vinoteca.controllers;
 
 import com.tequila.ecommerce.vinoteca.models.Product;
 import com.tequila.ecommerce.vinoteca.services.ProductServices;
-import javax.persistence.EntityNotFoundException;
-import javax.validation.Valid;
+import jakarta.persistence.EntityNotFoundException;
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
