@@ -1,6 +1,5 @@
 package com.tequila.ecommerce.vinoteca.controllers;
 
-
 import com.tequila.ecommerce.vinoteca.models.User;
 import com.tequila.ecommerce.vinoteca.services.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,37 +13,15 @@ import java.util.List;
 @RequestMapping("/usuario")
 public class UserController {
 
-
     @Autowired
     private UserService userService;
 
-    // GET /usuario
+    // Obtener todos los usuarios
     @GetMapping
-    public ResponseEntity<List<User>> obtenerTodosLosUsuarios(){
+    public ResponseEntity<List<User>> obtenerTodosLosUsuarios() {
         List<User> users = userService.obtenerTodosLosUsuarios();
         return new ResponseEntity<>(users, HttpStatus.OK);
     }
-    // GET /usuario/{id}
-    @GetMapping("/{id}")
-    public ResponseEntity<User> obtenerUsuariosPorId(@PathVariable Long id){
-        User user = userService.obtenerUsuarioPorId(id);
-        if(user!=null){
-            return new ResponseEntity<>(user, HttpStatus.OK);
-        }
-        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-    }
 
-    // POST /usuario
-    @PostMapping
-    public ResponseEntity<User> guardarUsuario(@RequestBody User user){
-        User nuevoUsuario = userService.guardarUsuario(user);
-        return new ResponseEntity<>(nuevoUsuario, HttpStatus.CREATED);
-    }
-
-    // DELETE /usuario/{id}
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> eliminarUsuario(@PathVariable Long id){
-        userService.eliminarUsuario(id);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
-    }
+    // Obtener usuario por ID
 }
