@@ -6,7 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ProductRepository extends JpaRepository<Product, Long> {
+public interface ProductRepository extends JpaRepository<Product, Long> {//Hereda métodos para realizar operaciones CRUD Product, 
+																		////donde el ID es de tipo Long.
     List<Product> findByTipoBebida(String tipoBebida);
-
+//findByTipoBebida Permite obtener una lista de productos filtrados por el tipo de bebida
 }
