@@ -3,5 +3,6 @@ package com.tequila.ecommerce.vinoteca.repository;
 import com.tequila.ecommerce.vinoteca.models.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository  extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Long> {
+    User findByEmail(String email);
 }
