@@ -7,25 +7,28 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-@Service
+@Service // Esto indica que es un componente de servicio de Spring
+		//responsable de contener la lógica de negocio y ser gestionado por el contenedor de Spring
 public class CategoryServices {
 
-    @Autowired
+    @Autowired //Inyección automática del CategoryRepository
     private CategoryRepository categoriaRepository;
 
-    public List<Category> obtenerTodasLasCategorias(){
+    public List<Category> obtenerTodasLasCategorias(){ //Devuelve una lista con todas las categorías de la bd.
         return categoriaRepository.findAll();
     }
-    public Category obtenerCategoriaPorId(Long id) {
+    public Category obtenerCategoriaPorId(Long id) {//Busca y retorna específica por su id. Si no la encuentra,  null
         return categoriaRepository.findById(id).orElse(null);
     }
 
-    public Category guardarCategorias(Category categoria) {
+    public Category guardarCategorias(Category categoria) {//Guarda una nueva categoría o actualiza una existente.
+
 
         return categoriaRepository.save(categoria);
     }
 
-    public void eliminarCategorias(Long id) {
+    public void eliminarCategorias(Long id) {//Elimina la categoría con el id especificado.
+
         categoriaRepository.deleteById(id);
     }
 
